Refactor IDynamicPerson to separate file. Add logic to handle parallel procession of get people.
