+++ src/components/mgt-people/mgt-people.ts
@@ -10,11 +10,11 @@ import { customElement, html, property, TemplateResult } from 'lit-element';
 import { repeat } from 'lit-html/directives/repeat';
 import { getPeople, getPeopleFromGroup } from '../../graph/graph.people';
 import { getUsersForPeopleQueries, getUsersForUserIds } from '../../graph/graph.user';
+import { IDynamicPerson } from '../../graph/Types';
 import { Providers } from '../../Providers';
 import { ProviderState } from '../../providers/IProvider';
 import '../../styles/fabric-icon-font';
 import { arraysAreEqual } from '../../utils/Utils';
-import { IDynamicPerson } from '../mgt-person/mgt-person';
 import { MgtTemplatedComponent } from '../templatedComponent';
 import { PersonCardInteraction } from './../PersonCardInteraction';
 import { styles } from './mgt-people-css';
