+++ src/components/mgt-person/mgt-person.ts
@@ -11,6 +11,7 @@ import { classMap } from 'lit-html/directives/class-map';
 import { findPerson, getEmailFromGraphEntity } from '../../graph/graph.people';
 import { getPersonImage } from '../../graph/graph.photos';
 import { getUserWithPhoto } from '../../graph/graph.user';
+import { IDynamicPerson } from '../../graph/Types';
 import { Providers } from '../../Providers';
 import { ProviderState } from '../../providers/IProvider';
 import '../../styles/fabric-icon-font';
@@ -20,23 +21,6 @@ import { MgtFlyout } from '../sub-components/mgt-flyout/mgt-flyout';
 import { MgtTemplatedComponent } from '../templatedComponent';
 import { PersonCardInteraction } from './../PersonCardInteraction';
 import { styles } from './mgt-person-css';
-
-/**
- * IDynamicPerson describes the person object we use throughout mgt-person,
- * which can be one of three similar Graph types.
- *
- * In addition, this custom type also defines the optional `personImage` property,
- * which is used to pass the image around to other components as part of the person object.
- */
-export type IDynamicPerson = (MicrosoftGraph.User | MicrosoftGraph.Person | MicrosoftGraph.Contact) & {
-  /**
-   * personDetails.personImage is a toolkit injected property to pass image between components
-   * an optimization to avoid fetching the image when unnecessary.
-   *
-   * @type {string}
-   */
-  personImage?: string;
-};
 
 /**
  * The person component is used to display a person or contact by using their photo, name, and/or email address.
@@ -220,7 +204,7 @@ export class MgtPerson extends MgtTemplatedComponent {
    */
   public render() {
     // Loading
-    if (this.isLoadingState) {
+    if (this.isLoadingState && !this.personDetails) {
       return this.renderLoading();
     }
 
@@ -463,8 +447,6 @@ export class MgtPerson extends MgtTemplatedComponent {
    */
   protected async loadState() {
     const provider = Providers.globalProvider;
-    const graph = provider.graph.forComponent(this);
-
     if (!provider || provider.state === ProviderState.Loading) {
       return;
     }
@@ -474,6 +456,7 @@ export class MgtPerson extends MgtTemplatedComponent {
       return;
     }
 
+    const graph = provider.graph.forComponent(this);
     if (this.personDetails) {
       // in some cases we might only have name or email, but need to find the image
       // use @ for the image value to search for an image
@@ -484,20 +467,14 @@ export class MgtPerson extends MgtTemplatedComponent {
           this.personImage = image;
         }
       }
-      return;
-    }
-
-    // Use userId or 'me' query to get the person and image
-    if (this.userId || this.personQuery === 'me') {
+    } else if (this.userId || this.personQuery === 'me') {
+      // Use userId or 'me' query to get the person and image
       const person = await getUserWithPhoto(graph, this.userId);
 
       this.personDetails = person;
       this.personImage = this.getImage();
-      return;
-    }
-
-    // Use the personQuery to find our person.
-    if (this.personQuery) {
+    } else if (this.personQuery) {
+      // Use the personQuery to find our person.
       const people = await findPerson(graph, this.personQuery);
 
       if (people && people.length) {
