+++ src/graph/graph.user.ts
@@ -6,11 +6,11 @@
  */
 
 import { User } from '@microsoft/microsoft-graph-types';
-import { IDynamicPerson } from '../components/mgt-person/mgt-person';
 import { IGraph } from '../IGraph';
 import { prepScopes } from '../utils/GraphHelpers';
 import { findPerson } from './graph.people';
 import { getPersonImage } from './graph.photos';
+import { IDynamicPerson } from './Types';
 
 /**
  * async promise, returns Graph User data relating to the user logged in
@@ -123,20 +123,44 @@ export async function getUsersForPeopleQueries(graph: IGraph, peopleQueries: str
     return [];
   }
 
+  const batch = graph.createBatch();
   const people = [];
+
   for (const personQuery of peopleQueries) {
     if (personQuery !== '') {
-      const person = (await findPerson(graph, personQuery)) as IDynamicPerson[];
-      if (person && person.length) {
-        people.push(person[0]);
-
-        const image = await getPersonImage(graph, person[0]);
-        if (image) {
-          person[0].personImage = image;
-        }
-      }
+      batch.get(personQuery, `/me/people?$search="${personQuery}"`, ['people.read']);
     }
   }
 
-  return people;
+  try {
+    const response = await batch.execute();
+
+    for (const personQuery of peopleQueries) {
+      const person = response[personQuery];
+      if (person) {
+        people.push(person.value[0]);
+      }
+    }
+
+    return people;
+  } catch (_) {
+    try {
+      for (const personQuery of peopleQueries) {
+        if (personQuery !== '') {
+          const person = (await findPerson(graph, personQuery)) as IDynamicPerson[];
+          if (person && person.length) {
+            people.push(person[0]);
+
+            const image = await getPersonImage(graph, person[0]);
+            if (image) {
+              person[0].personImage = image;
+            }
+          }
+        }
+      }
+      return people;
+    } catch (_) {
+      return [];
+    }
+  }
 }
